moved opensha_task model into custom/rupture_generation
