+++ graphql_api/schema/__init__.py
@@ -2,7 +2,7 @@
 Module entry point
 """
 from graphql_api.schema.schema import root_schema
-from graphql_api.schema.opensha_task import (RuptureGenerationTask,
+from graphql_api.schema.custom.rupture_generation import (RuptureGenerationTask,
     RuptureGenerationTaskConnection,
     CreateRuptureGenerationTask)
