+++ graphql_api/schema/custom/rupture_generation.py
